Split dependency.py into two modules
This makes the hierarchy much clearer.

dependencies <- requirements <- resolvers
