+++ docs/requirementslib.models.dependencies.rst
@@ -1,0 +1,7 @@
+requirementslib.models.dependencies module
+========================================
+
+.. automodule:: requirementslib.models.dependencies
+    :members:
+    :undoc-members:
+    :show-inheritance:
