+++ src/requirementslib/models/__init__.py
@@ -2,10 +2,10 @@
 from __future__ import absolute_import
 
 
-__all__ = ["Requirement", "Lockfile", "Pipfile", "RequirementError", "DependencyResolver"]
+__all__ = ["Requirement", "Lockfile", "Pipfile", "DependencyResolver"]
 
 
 from .requirements import Requirement
 from .lockfile import Lockfile
 from .pipfile import Pipfile
-from .dependency import DependencyResolver
+from .resolvers import DependencyResolver
