+++ src/requirementslib/models/dependencies.py
@@ -1,49 +1,70 @@
 # -*- coding=utf-8 -*-
-import attr
+
 import copy
 import functools
+import os
+
+import attr
+import first
+import packaging.version
 import requests
-import six
-from first import first
-from packaging.version import parse as parse_version
-from .utils import (
-    full_groupby,
-    key_from_ireq,
-    is_pinned_requirement,
-    name_from_req,
-    make_install_requirement,
-    format_requirement,
-    version_from_ireq,
-    partialclass,
-)
-from .cache import CACHE_DIR, DependencyCache
+
 from .._compat import (
-    RequirementPreparer,
-    Resolver,
-    WheelCache,
-    RequirementSet,
-    PackageFinder,
-    TemporaryDirectory,
-    InstallRequirement,
-    FormatControl,
-    RequirementTracker,
     pip_version,
+    FormatControl,
+    InstallRequirement,
+    PackageFinder,
+    RequirementPreparer,
+    RequirementSet,
+    RequirementTracker,
+    Resolver,
+    TemporaryDirectory,
+    WheelCache,
 )
 from ..utils import (
     fs_str,
-    prepare_pip_source_args,
     get_pip_command,
-    temp_cd,
-    log,
     mkdir_p,
+    prepare_pip_source_args,
+    temp_cd,
+)
+from .cache import CACHE_DIR, DependencyCache
+from .utils import (
+    format_requirement,
+    full_groupby,
+    is_pinned_requirement,
+    key_from_ireq,
+    make_install_requirement,
+    name_from_req,
+    partialclass,
+    version_from_ireq,
 )
 
-import os
 
-DOWNLOAD_DIR = fs_str(os.path.join(CACHE_DIR, "pkgs"))
+PKGS_DOWNLOAD_DIR = fs_str(os.path.join(CACHE_DIR, "pkgs"))
 WHEEL_DOWNLOAD_DIR = fs_str(os.path.join(CACHE_DIR, "wheels"))
-DEPCACHE = DependencyCache()
+
+DEPENDENCY_CACHE = DependencyCache()
 WHEEL_CACHE = WheelCache(CACHE_DIR, FormatControl(None, None))
+
+
+def find_all_matches(finder, ireq):
+    """Find all matching dependencies using the supplied finder and the
+    given ireq.
+
+    :param finder: A package finder for discovering matching candidates.
+    :type finder: :class:`~pip._internal.index.PackageFinder`
+    :param ireq: An install requirement.
+    :type ireq: :class:`~pip._internal.req.req_install.InstallRequirement`
+    :return: A list of matching candidates.
+    :rtype: list[:class:`~pip._internal.index.InstallationCandidate`]
+    """
+
+    candidates = set(finder.find_all_candidates(ireq.name))
+    if not ireq.specifier:
+        return list(candidates)
+    matches = [m for m in candidates if m.version in ireq.specifier]
+    return matches
 
 
 @attr.s
@@ -64,7 +85,7 @@ class AbstractDependency(object):
         :rtype: set(str)
         """
 
-        return set(parse_version(version_from_ireq(c)) for c in self.candidates)
+        return set(packaging.version.parse(version_from_ireq(c)) for c in self.candidates)
 
     def compatible_versions(self, other):
         """Find compatible version numbers between this abstract
@@ -99,7 +120,7 @@ class AbstractDependency(object):
         candidates = [
             c
             for c in self.candidates
-            if parse_version(version_from_ireq(c)) in compatible_versions
+            if packaging.version.parse(version_from_ireq(c)) in compatible_versions
         ]
         dep_dict = {}
         candidate_strings = [format_requirement(c) for c in candidates]
@@ -161,7 +182,7 @@ class AbstractDependency(object):
         else:
             candidates = [requirement.ireq]
         candidates = sorted(
-            set(candidates), key=lambda k: parse_version(version_from_ireq(k)),
+            set(candidates), key=lambda k: packaging.version.parse(version_from_ireq(k)),
         )
         return cls(
             name=name,
@@ -179,219 +200,6 @@ class AbstractDependency(object):
         req = Requirement.from_line(line)
         abstract_dep = cls.from_requirement(req, parent=parent)
         return abstract_dep
-
-
-class ResolutionError(Exception):
-    pass
-
-
-@attr.s
-class DependencyResolver(object):
-    pinned_deps = attr.ib(default=attr.Factory(dict))
-    #: A dictionary of abstract dependencies by name
-    dep_dict = attr.ib(default=attr.Factory(dict))
-    #: A dictionary of sets of version numbers that are valid for a candidate currently
-    candidate_dict = attr.ib(default=attr.Factory(dict))
-    #: A historical record of pins
-    pin_history = attr.ib(default=attr.Factory(dict))
-
-    @property
-    def dependencies(self):
-        return list(self.dep_dict.values())
-
-    @property
-    def resolution(self):
-        return list(self.pinned_deps.values())
-
-    def add_abstract_dep(self, dep):
-        """Add an abstract dependency by either creating a new entry or
-        merging with an old one.
-
-        :param dep: An abstract dependency to add
-        :type dep: :class:`~requirementslib.models.dependency.AbstractDependency`
-        :raises ResolutionError: Raised when the given dependency is not compatible with
-                                 an existing abstract dependency.
-        """
-
-        if dep.name in self.dep_dict:
-            compatible_versions = self.dep_dict[dep.name].compatible_versions(dep)
-            if compatible_versions:
-                self.candidate_dict[dep.name] = compatible_versions
-                self.dep_dict[dep.name] = self.dep_dict[
-                    dep.name
-                ].compatible_abstract_dep(dep)
-            else:
-                raise ResolutionError
-        else:
-            self.candidate_dict[dep.name] = dep.version_set
-            self.dep_dict[dep.name] = dep
-
-    def pin_deps(self):
-        """Pins the current abstract dependencies and adds them to the history dict.
-
-        Adds any new dependencies to the abstract dependencies already present by
-        merging them together to form new, compatible abstract dependencies.
-        """
-
-        for name in list(self.dep_dict.keys()):
-            candidates = self.dep_dict[name].candidates[:]
-            abs_dep = self.dep_dict[name]
-            while candidates:
-                pin = candidates.pop()
-                new_version = version_from_ireq(pin)
-                # Move on from existing pins if the new pin isn't compatible
-                if name in self.pinned_deps:
-                    old_version = version_from_ireq(self.pinned_deps[name])
-                    if (new_version != old_version and
-                            new_version not in self.candidate_dict[name]):
-                        continue
-                pin.parent = abs_dep.parent
-                pin_subdeps = self.dep_dict[name].get_deps(pin)
-                backup = self.dep_dict.copy(), self.candidate_dict.copy()
-                try:
-                    for pin_dep in pin_subdeps:
-                        self.add_abstract_dep(pin_dep)
-                except ResolutionError:
-                    self.dep_dict, self.candidate_dict = backup
-                    continue
-                else:
-                    self.pinned_deps[name] = pin
-                    break
-
-    def resolve(self, root_nodes, max_rounds=20):
-        """Resolves dependencies using a backtracking resolver and multiple endpoints.
-
-        Note: this resolver caches aggressively.
-        Runs for *max_rounds* or until any two pinning rounds yield the same outcome.
-
-        :param root_nodes: A list of the root requirements.
-        :type root_nodes: list[:class:`~requirementslib.models.requirements.Requirement`]
-        :param max_rounds: The max number of resolution rounds, defaults to 20
-        :param max_rounds: int, optional
-        :raises RuntimeError: Raised when max rounds is exceeded without a resolution.
-        """
-        if self.dep_dict:
-            raise RuntimeError("Do not use the same resolver more than once")
-
-        # Coerce input into AbstractDependency instances.
-        # We accept str, Requirement, and AbstractDependency as input.
-        for dep in root_nodes:
-            if isinstance(dep, six.string_types):
-                dep = AbstractDependency.from_string(dep)
-            elif not isinstance(dep, AbstractDependency):
-                dep = AbstractDependency.from_requirement(dep)
-            self.add_abstract_dep(dep)
-
-        for round_ in range(max_rounds):
-            self.pin_deps()
-            self.pin_history[round_] = self.pinned_deps.copy()
-
-            if round_ > 0:
-                previous_round = set(self.pin_history[round_ - 1].values())
-                current_values = set(self.pin_history[round_].values())
-                difference = current_values - previous_round
-            else:
-                difference = set(self.pin_history[round_].values())
-
-            log.debug("\n")
-            log.debug("{:=^30}".format(" Round {0} ".format(round_)))
-            log.debug("\n")
-            if difference:
-                log.debug("New Packages: ")
-                for d in difference:
-                    log.debug("{:>30}".format(format_requirement(d)))
-            elif round_ >= 3:
-                log.debug("Stable Pins: ")
-                for d in current_values:
-                    log.debug("{:>30}".format(format_requirement(d)))
-                return
-            else:
-                log.debug("No New Packages.")
-        # TODO: Raise a better error.
-        raise RuntimeError("cannot resolve after {} rounds".format(max_rounds))
-
-
-def get_resolver(sources=None):
-    """Given a list of sources, return a finder, preparer, and resolver.
-
-    :param sources: A list of pipfile-formatted sources, defaults to None
-    :param sources: list[dict], optional
-    :return: A 3-tuple of finder, preparer, resolver
-    :rtype: (:class:`~pip._internal.index.PackageFinder`, :class:`~pip._internal.operations.prepare.RequirementPreparer`, :class:`~pip._internal.resolve.Resolver`)
-    """
-
-    pip_command = get_pip_command()
-    if not sources:
-        sources = [
-            {"url": "https://pypi.org/simple", "name": "pypi", "verify_ssl": True}
-        ]
-    pip_args = []
-    pip_args = prepare_pip_source_args(sources, pip_args)
-    pip_options, _ = pip_command.parser.parse_args(pip_args)
-    pip_options.cache_dir = CACHE_DIR
-    mkdir_p(CACHE_DIR)
-    session = pip_command._build_session(pip_options)
-    wheel_cache = WheelCache(CACHE_DIR, pip_options.format_control)
-    finder = PackageFinder(
-        find_links=[],
-        index_urls=[s.get("url") for s in sources],
-        trusted_hosts=[],
-        allow_all_prereleases=pip_options.pre,
-        session=session,
-    )
-    download_dir = DOWNLOAD_DIR
-    mkdir_p(download_dir)
-    _build_dir = TemporaryDirectory(fs_str("build"))
-    _source_dir = TemporaryDirectory(fs_str("source"))
-    preparer = partialclass(
-        RequirementPreparer,
-        build_dir=_build_dir.name,
-        src_dir=_source_dir.name,
-        download_dir=download_dir,
-        wheel_download_dir=WHEEL_DOWNLOAD_DIR,
-        progress_bar="off",
-        build_isolation=False,
-    )
-    resolver = partialclass(
-        Resolver,
-        finder=finder,
-        session=session,
-        upgrade_strategy="to-satisfy-only",
-        force_reinstall=True,
-        ignore_dependencies=False,
-        ignore_requires_python=True,
-        ignore_installed=True,
-        isolated=False,
-        wheel_cache=wheel_cache,
-        use_user_site=False,
-    )
-    if parse_version(pip_version) >= parse_version('18.0.0'):
-        with RequirementTracker() as req_tracker:
-            preparer = preparer(req_tracker=req_tracker)
-            resolver = resolver(preparer=preparer)
-    else:
-        preparer = preparer()
-        resolver = resolver(preparer=preparer)
-    return finder, preparer, resolver
-
-
-def find_all_matches(finder, ireq):
-    """Find all matching dependencies using the supplied finder and the
-    given ireq.
-
-    :param finder: A package finder for discovering matching candidates.
-    :type finder: :class:`~pip._internal.index.PackageFinder`
-    :param ireq: An install requirement.
-    :type ireq: :class:`~pip._internal.req.req_install.InstallRequirement`
-    :return: A list of matching candidates.
-    :rtype: list[:class:`~pip._internal.index.InstallationCandidate`]
-    """
-
-    candidates = set(finder.find_all_candidates(ireq.name))
-    if not ireq.specifier:
-        return list(candidates)
-    matches = [m for m in candidates if m.version in ireq.specifier]
-    return matches
 
 
 def get_abstract_dependencies(reqs, sources=None, parent=None):
@@ -476,8 +284,8 @@ def get_dependencies_from_wheel_cache(ireq):
     matches = WHEEL_CACHE.get(ireq.link, name_from_req(ireq.req))
     if matches:
         matches = set(matches)
-        if not DEPCACHE.get(ireq):
-            DEPCACHE[ireq] = [format_requirement(m) for m in matches]
+        if not DEPENDENCY_CACHE.get(ireq):
+            DEPENDENCY_CACHE[ireq] = [format_requirement(m) for m in matches]
         return matches
     return
 
@@ -510,8 +318,8 @@ def get_dependencies_from_json(ireq):
                 # TODO: Get dependencies for matching extra.
                 yield format_requirement(i)
 
-    if ireq not in DEPCACHE:
-        reqs = DEPCACHE[ireq] = list(gen(ireq))
+    if ireq not in DEPENDENCY_CACHE:
+        reqs = DEPENDENCY_CACHE[ireq] = list(gen(ireq))
         req_iter = iter(reqs)
     else:
         req_iter = gen(ireq)
@@ -527,8 +335,8 @@ def get_dependencies_from_cache(dep):
     :rtype: set(str) or None
     """
 
-    if dep in DEPCACHE:
-        return set(DEPCACHE[dep])
+    if dep in DEPENDENCY_CACHE:
+        return set(DEPENDENCY_CACHE[dep])
     return
 
 
@@ -563,8 +371,72 @@ def get_dependencies_from_index(dep, sources=None):
     requirements = reqset.requirements.values()
     reqset.cleanup_files()
     reqs = set(requirements)
-    DEPCACHE[dep] = [format_requirement(r) for r in reqs]
+    DEPENDENCY_CACHE[dep] = [format_requirement(r) for r in reqs]
     return reqs
+
+
+def get_resolver(sources=None):
+    """Given a list of sources, return a finder, preparer, and resolver.
+
+    :param sources: A list of pipfile-formatted sources, defaults to None
+    :param sources: list[dict], optional
+    :return: A 3-tuple of finder, preparer, resolver
+    :rtype: (:class:`~pip._internal.index.PackageFinder`, :class:`~pip._internal.operations.prepare.RequirementPreparer`, :class:`~pip._internal.resolve.Resolver`)
+    """
+
+    pip_command = get_pip_command()
+    if not sources:
+        sources = [
+            {"url": "https://pypi.org/simple", "name": "pypi", "verify_ssl": True}
+        ]
+    pip_args = []
+    pip_args = prepare_pip_source_args(sources, pip_args)
+    pip_options, _ = pip_command.parser.parse_args(pip_args)
+    pip_options.cache_dir = CACHE_DIR
+    mkdir_p(CACHE_DIR)
+    session = pip_command._build_session(pip_options)
+    wheel_cache = WheelCache(CACHE_DIR, pip_options.format_control)
+    finder = PackageFinder(
+        find_links=[],
+        index_urls=[s.get("url") for s in sources],
+        trusted_hosts=[],
+        allow_all_prereleases=pip_options.pre,
+        session=session,
+    )
+    download_dir = PKGS_DOWNLOAD_DIR
+    mkdir_p(download_dir)
+    _build_dir = TemporaryDirectory(fs_str("build"))
+    _source_dir = TemporaryDirectory(fs_str("source"))
+    preparer = partialclass(
+        RequirementPreparer,
+        build_dir=_build_dir.name,
+        src_dir=_source_dir.name,
+        download_dir=download_dir,
+        wheel_download_dir=WHEEL_DOWNLOAD_DIR,
+        progress_bar="off",
+        build_isolation=False,
+    )
+    resolver = partialclass(
+        Resolver,
+        finder=finder,
+        session=session,
+        upgrade_strategy="to-satisfy-only",
+        force_reinstall=True,
+        ignore_dependencies=False,
+        ignore_requires_python=True,
+        ignore_installed=True,
+        isolated=False,
+        wheel_cache=wheel_cache,
+        use_user_site=False,
+    )
+    if packaging.version.parse(pip_version) >= packaging.version.parse('18'):
+        with RequirementTracker() as req_tracker:
+            preparer = preparer(req_tracker=req_tracker)
+            resolver = resolver(preparer=preparer)
+    else:
+        preparer = preparer()
+        resolver = resolver(preparer=preparer)
+    return finder, preparer, resolver
 
 
 def get_grouped_dependencies(constraints):
@@ -575,7 +447,7 @@ def get_grouped_dependencies(constraints):
     # versions by popping them off of a stack and checking for the conflicting package
     for _, ireqs in full_groupby(constraints, key=key_from_ireq):
         ireqs = list(ireqs)
-        editable_ireq = first(ireqs, key=lambda ireq: ireq.editable)
+        editable_ireq = first.first(ireqs, key=lambda ireq: ireq.editable)
         if editable_ireq:
             yield editable_ireq  # ignore all the other specs: the editable one is the one that counts
             continue
